+++ CHANGELOG.md
@@ -11,7 +11,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## TODO
 - SimpleCache closeable
-- DatagramPacket -> Scion
 
 - Remove Path from ScionSocketAddress?
   - Make Path subclass of ScionSocketAddress?
@@ -186,6 +185,8 @@ TODO
   (it did that despite the javadoc saying otherwise).
   [#275](https://github.com/scionproto-contrib/jpan/pull/275)
   [#277](https://github.com/scionproto-contrib/jpan/pull/277)
+- Added ScionDatagram packet
+  [#278](https://github.com/scionproto-contrib/jpan/pull/278)
 
 ### Fixed